+++ translations/snaptray_el.ts
@@ -3438,6 +3438,10 @@ File size: %3</source>
 <context>
     <name>PinToolOptionsViewModel</name>
     <message>
+        <source>Scroll here to adjust the mosaic brush size</source>
+        <translation>Κάντε κύλιση εδώ για να προσαρμόσετε το μέγεθος του πινέλου μωσαϊκού</translation>
+    </message>
+    <message>
         <source>Automatically detect and blur faces and credentials</source>
         <translation>Αυτόματος εντοπισμός και θόλωση προσώπων και διαπιστευτηρίων</translation>
     </message>