+++ translations/snaptray_es_MX.ts
@@ -3444,6 +3444,10 @@ Tamaño del archivo: %3</translation>
 <context>
     <name>PinToolOptionsViewModel</name>
     <message>
+        <source>Scroll here to adjust the mosaic brush size</source>
+        <translation>Desplázate aquí para ajustar el tamaño del pincel de mosaico</translation>
+    </message>
+    <message>
         <source>Automatically detect and blur faces and credentials</source>
         <translation>Detectar y difuminar automáticamente rostros y credenciales</translation>
     </message>