+++ translations/snaptray_lt.ts
@@ -3438,6 +3438,10 @@ Failo dydis: %3</translation>
 <context>
     <name>PinToolOptionsViewModel</name>
     <message>
+        <source>Scroll here to adjust the mosaic brush size</source>
+        <translation>Slinkite čia, kad pakeistumėte mozaikos teptuko dydį</translation>
+    </message>
+    <message>
         <source>Automatically detect and blur faces and credentials</source>
         <translation>Automatiškai aptikti ir sulieti veidus bei prisijungimo duomenis</translation>
     </message>