+++ translations/snaptray_pt_PT.ts
@@ -3438,6 +3438,10 @@ Tamanho do ficheiro: %3</translation>
 <context>
     <name>PinToolOptionsViewModel</name>
     <message>
+        <source>Scroll here to adjust the mosaic brush size</source>
+        <translation>Desloque aqui para ajustar o tamanho do pincel de mosaico</translation>
+    </message>
+    <message>
         <source>Automatically detect and blur faces and credentials</source>
         <translation>Detetar e desfocar automaticamente rostos e credenciais</translation>
     </message>