+++ translations/snaptray_th.ts
@@ -3199,6 +3199,10 @@ Password: %2</source>
 <context>
     <name>PinToolOptionsViewModel</name>
     <message>
+        <source>Scroll here to adjust the mosaic brush size</source>
+        <translation>เลื่อนที่นี่เพื่อปรับขนาดแปรงโมเสก</translation>
+    </message>
+    <message>
         <source>Automatically detect and blur faces and credentials</source>
         <translation>ตรวจจับและเบลอใบหน้าและข้อมูลประจำตัวโดยอัตโนมัติ</translation>
     </message>